Add rough draft of deme.cc and deme.hh
None of the functions have an actual random number generator yet. I added comments where there needs to be one, along with the range of numbers it should select from. I implemented the roulette wheel selection linked in the assignment description. To make the computation of a chromosome's probability smoother, I added a data member that's the sum fitness of the entire deme. This gets computed at construction and at the end of ever generation. All other functions are implemented as described in the assignment (probably, unless my code is buggy).